Remove Unecessary File From App Generation
When the app generator was run in with `--coffee` there was a router.js file
being created in the repo. It looks like the router.js file is being generated
through the router/template rather than the app/template and there is no
coffeecript option for it. This was the simplest fix given the the router is not
strictly necessary for the application that is generated.

I also added an assertion to demonstrate that the router.js file wasn't created.